pyproject.toml
[tool.poetry]
name = "update_jira_issues"
version = "0.1.0"
description = ""
authors = ["kimbyeongcheol <byeongcheol.kim@wesang.com>"]

[tool.poetry.dependencies]
python = "^3.10"
jira = "^3.5.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
